Separate accordion data into its own file
